feat: putting profile info onto separate component ProfileUserInfo
